pyproject.toml
[tool.poetry]
name = "cnv-detection"
version = "0.1.0"
description = "Detection of CNV variants using ML/DL methods"
authors = ["Bartosz Lewandowski <b.lewandowski06@gmail.com"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.4.3"
numpy = "^1.23.1"
biopython = "^1.79"
pybedtools = "^0.9.0"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
black = "^22.6.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
